Feature: Adding webhook notification file this change will include notification.php that support webhooks request our plugin endpoint to be able to change orders status
Why is this change neccesary?

this change is necesary because order statuses are not changing in a common order life cycle

How does it address the issue?

What side effects does this change have?
this change maybe affect the upload to marketplace, the custom validation of marketplace
rejet files like this