Remove unused Credit Card class and update Conekta SPEI payment emails
This commit removes the unused Card class from the Conekta library and updates the templates for SPEI payment emails in several different language versions. This step is essential in keeping the codebase neat and manageable by eradicating obsolete parts. Additionally, updating the email templates ensures that communication with customers is well-maintained and relevant to their needs and expectations.