Add Makefile for zipping project files
This commit adds a new Makefile that includes a "zip" command. The "zip" command creates a zip file of the project, excluding certain files and directories specified by the exclusion patterns. The ".PHONY" target is also added to ensure that the "zip" command is always executed, regardless of whether there are any files with matching names.